requirements.txt
flash_attn==2.7.4.post1
torch==2.6.0
transformers==4.48.2
accelerate==1.3.0
soundfile==0.13.1
pillow==11.1.0
scipy==1.15.2
torchvision==0.21.0
backoff==2.2.1
peft==0.15.0
diffusers @ git+https://github.com/huggingface/diffusers@main
datasets==3.6.0